Sprzątanie, PESEL, nr tel
Dodany został plik do generowania peselu, tak by był poprawny, zawierał cyfry kontrolne itp;
Dodane zostało generowanie numerów telefonu, które pobiera początek z pliku prefiksy.csv;
Porodzielane zostały wszystkie generatory z pliku main na osobne by było czytelniej;
