+++ plec_generator.py
@@ -1,0 +1,4 @@
+import random
+
+def generuj_plec():
+    return random.choice(["M", "K"])
